.float()

+++ imagenet-2012/data_load.py
@@ -187,7 +187,7 @@ class ToTensor(object):
         image = image.transpose((2, 0, 1))
 
         return {
-            'image': torch.from_numpy(image),
+            'image': torch.from_numpy(image).float(),
             'annotation': annotation,
         }
